pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[project]
name = "labelforge"
version = "0.1.0"
description = "AI-driven export labeling automation for Nakoda Art & Craft"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "redis>=5.0.0",
    "boto3>=1.34.0",
    "httpx>=0.27.0",
    "reportlab>=4.0.0",
    "pypdf>=4.0.0",
    "pymupdf>=1.23.0",
    "openpyxl>=3.1.0",
    "Pillow>=10.0.0",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "openai>=1.30.0",
    "structlog>=24.1.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp-proto-http>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.48b0",
    "opentelemetry-instrumentation-httpx>=0.48b0",
    "opentelemetry-instrumentation-logging>=0.48b0",
    "prometheus-client>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-redis>=4.6.0",
    "types-boto3>=1.34.0",
    "types-passlib>=1.7.7",
    "types-python-jose>=3.3.4",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["labelforge"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
allow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "reportlab.*",
    "fitz.*",
    "PIL.*",
    "pymupdf.*",
    "temporalio.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring external services",
    "asyncio: marks tests as async",
    "stress: marks multi-tenancy / concurrency / load stress tests (opt-in)",
    "postgres: marks tests that require a real PostgreSQL backend",
]
